Fix deallocation of db_conn_t
The free() at the end of db_init_conn_nif() is wrong. There is
no temporary object that needs to be deallocated.


Maybe this would be even easier to understand by introducing a struct
for the resource object that holds the pointer to db_conn_t:

typedef struct {
  db_conn_t *conn;
} db_conn_resource;

That way you would avoid the double pointer syntax and the destructor
could be:

void db_res_destructor(ErlNifEnv *env, void *res) {
  db_conn_resource *conn_res = (db_conn_resource*) res;
  db_free_conn(conn_res->conn);
}